hemos añadido foto de empleado existente

+++ src/main/java/com/example/controller/EmpleadoController.java
@@ -155,13 +155,17 @@ public class EmpleadoController {
 		}
 		
         //Antes de persistir el empleado hay que eliminar los telefonos y los correos que tenga
-		if (telefonoService.existsByEmpleado(empleado))
-		    telefonoService.deleteByEmpleado(empleado);
+		if (empleado.getId() != 0) {
+			
+			if (telefonoService.existsByEmpleado(empleado))
+				telefonoService.deleteByEmpleado(empleado);
 
-		if (correoService.existsByEmpleado(empleado))
-			correoService.deleteByEmpleado(empleado);
-
-
+			if (correoService.existsByEmpleado(empleado))
+				correoService.deleteByEmpleado(empleado);
+		}
+		
+		
+		
 
 		// Se recibe un objeto Empleado con los datos del formulario
 		// Se envía a la capa de servicios para que lo guarde en la BD
